requirements.txt
torch
numpy
pillow
mlflow
streamlit
matplotlib